Deprecate shrinking OpenMP team size to max
Now, if a user asks for a team size bigger
than the max, an exception will be thrown.
The unit tests that failed with this change
were updated to check for the max team size.
Right now that is less than nice because the
functor in question is needed to check the
max allowable team size...